format output

+++ package/zsh/install.sh
@@ -2,10 +2,9 @@
 
 source $LIB
 
-echo
-info "==>" "config zsh"
+brewinstall zsh
 
-echo "==> install oh my zsh"
+info "==>" "install oh my zsh"
 if [ ! -d ~/.oh-my-zsh ]; then
   /bin/sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh)"
 else
@@ -15,12 +14,20 @@ fi
 # install customized zsh theme
 SOURCE_FILE=$INSTALL_HOME/package/zsh/simple.zsh-theme
 TARGET_FILE=~/.oh-my-zsh/custom/themes/simple.zsh-theme
-backup $TARGET_FILE
-echo "==>" "install customized zsh theme"
-cp -p $SOURCE_FILE $TARGET_FILE
-log "copy: $SOURCE_FILE -> $TARGET_FILE"
+if [ ! -f "$TARGET_FILE" ]; then
+  msg="install"
+else
+  msg="re-install"
+fi
+read -r -p "do you want to $msg simple.zsh-theme (y/n): " ans
+if [ $ans = "y" ]; then
+  backup $TARGET_FILE
+  info "==>" "install simple.zsh-theme"
+  cp -p $SOURCE_FILE $TARGET_FILE
+  log "copy: $SOURCE_FILE -> $TARGET_FILE"
+fi
 
 # symlink conf
-echo "==>" "config zshrc"
+info "==>" "config zshrc"
 backup ~/.zshrc
 linkconf $INSTALL_HOME/package/zsh/zshrc ~/.zshrc
